Adjust auth domain: just provide authenticate & issue token
