+++ domain/src/main/java/hk/onedegree/domain/auth/constant/Constant.java
@@ -1,0 +1,5 @@
+package hk.onedegree.domain.auth.constant;
+
+public class Constant {
+    final public static String issuer = "cymetrics";
+}
