+++ domain/src/main/java/hk/onedegree/domain/auth/repository/UserRepository.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 
 public interface UserRepository {
     public Optional<User> findByEmail(String email);
+    public Optional<User> findById(String id);
     public void save(User user);
     public void delete(User user);
  }
