+++ persistence/src/main/java/hk/onedegree/persistence/MemUserRepository.java
@@ -8,6 +8,7 @@ import java.util.concurrent.ConcurrentHashMap;
 
 public class MemUserRepository implements UserRepository {
     ConcurrentHashMap<String, User> emailUserMap = new ConcurrentHashMap<>();
+    ConcurrentHashMap<String, User> idUserMap = new ConcurrentHashMap<>();
 
     private static MemUserRepository instance = new MemUserRepository();
     public static MemUserRepository getInstance(){
@@ -20,12 +21,19 @@ public class MemUserRepository implements UserRepository {
     }
 
     @Override
+    public Optional<User> findById(String id) {
+        return Optional.ofNullable(this.idUserMap.get(id));
+    }
+
+    @Override
     public void save(User user) {
         this.emailUserMap.put(user.getEmail(), user);
+        this.idUserMap.put(user.getId(), user);
     }
 
     @Override
     public void delete(User user) {
         this.emailUserMap.remove(user.getEmail());
+        this.idUserMap.remove(user.getEmail());
     }
 }
