+++ pom.xml
@@ -37,6 +37,7 @@
         <assertj-core.version>3.19.0</assertj-core.version>
         <logback.version>1.2.3</logback.version>
         <jsonassert.version>1.5.0</jsonassert.version>
+        <mockito-core.version>3.8.0</mockito-core.version>
     </properties>
 
     <dependencies>
@@ -97,6 +98,19 @@
             <groupId>org.skyscreamer</groupId>
             <artifactId>jsonassert</artifactId>
             <version>${jsonassert.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <!-- https://mvnrepository.com/artifact/org.mockito/mockito-core -->
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-core</artifactId>
+            <version>${mockito-core.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-junit-jupiter</artifactId>
+            <version>${mockito-core.version}</version>
             <scope>test</scope>
         </dependency>
